requirements.txt
websocket-client==1.7.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
